requirements.txt
# Core
gensim>=4.3.0
fasttext-wheel>=0.9.2     

# Numerics
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3

# Utils
tqdm>=4.65.0